Reword labels for the bottom buttons of the top bar menu
I think it makes sense to call it the Freon wiki rather than the GitHub
wiki, because code hosting could change, and this wiki is specifically
all about Freon. And I think adding the "Go to" phrase both clarifies
what the button does, and also sorta invites the user to actually click
on the button.

And as for "Sensors Settings", this might be a little nit-picky, but you
don't *usually* pluralize a noun when it's acting as a modifier on
another noun.