Visualize hot (on panel) sensors that became stale
When a hwmon module was unloaded or crashed any sensors it had that were
also pinned to the panel (i.e. a 'hot sensor') retained the value
they had last, hiding the failure.

To fix that, this first clears all hot labels, replacing them with a
placeholder, before checking the new sensor data.  The placeholder
chosen is ⚠ (\u26a0 warning sign), after some experimentation with other
alternatives.

For consistency, the placeholder used on initialization is also
changed.